backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
channels
channels-redis
daphne